requirements.txt
slack-sdk==3.21.3
aiohttp
python-dotenv==1.0.0